Update patch set 3
Patch Set 3:

(1 comment)

This fixes the thickness errors that occur in the recently added GM, but only in non-MSAA. No other GMs change. The problem was that when the stroke was thin we were interpolating a non-bloated rect across a bloated rect and using that interpolated value as the device space distance to the center of the stroke.

Patch-set: 3
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5085 <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>