Update patch set 2
Patch Set 2:

Looks good, but might want to add a gm, or some other to not accidentally slip back into our old ways down the road.

possible gm to land prior to this "fix":
- make LCD capable offscreen canvas
- saveLayer (but triggering non-lcd geometry somehow)
- saveLayer again (preserving the request for lcd) and then draw LCD text

You could draw a zoomed-up version of the offscreen into the gm, so reviewers can easily see the striping of the LCD in the pre-fix version.

Patch-set: 2
CC: Gerrit User 5649 <5649@5086850b-e599-37d2-8ec8-ff16aeccbf02>