Update patch set 1
Patch Set 1:

This gm doesn't really test anything anymore with this change! (Except that we can make our own subpixel masks from paths, which is already tested lots of other ways.) On macOS requesting LCD is different. On 10.13 and earlier we expect it to create subpixel masks, on 10.14 and later we expect it to make the glyphs a bit bolder if hinting is set to normal.

That being said, this particular gm is probably completely superseded by the typefacerendering gm, except fo the size part, and this should probably have more sizes since dflcd exists now.

Patch-set: 1
Reviewer: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>