Update patch set 1
Patch Set 1:

It's testing that it draws 'like' lcd. It's different from not requesting it. On the other hand I think textblobmixedsizes and textblobmixedsizes_df probably do a much better job of testing what this was trying to test, and typefacerendering does the rest of what this was trying to do better. So yeah, can probably just kill this one.

Though note that those two also will reflect this 'requested subpixel edging' doesn't actually give subpixel edging but does change they way that the glyph are drawn (vs not requesting subpixel edging) on macOS. But it won't be quite as obvious since there's a lot more going on in those.

Patch-set: 1
Reviewer: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>