Label plan
I forgot about how this would interact with instructions that have
displacements not right at the end of the instruction,
(imm_byte_after_operand) and I think it'd be more complex to get this
plan working with those than doing this is worth.  Uploading my partial
work in case I change my mind.

Change-Id: I4a72ba672bc9b04bfa5b4ada6a2f344ce0d47233