Update patch set 1
Patch Set 1:

It's taken me a while to think through this, but I think it can work, using the displacement payloads themselves as a linked list to point back to prior instructions interested in where a label ends up.  Will try implementing this tomorrow but I thought you might be interested in seeing the plan early.

Patch-set: 1
CC: Gerrit User 5210 <5210@5086850b-e599-37d2-8ec8-ff16aeccbf02>
CC: Gerrit User 5649 <5649@5086850b-e599-37d2-8ec8-ff16aeccbf02>