only pass a real paint to drawPicture if we need to
If the caller passed nullptr for the picture's paint, then they did not
intend for the picture to be drawn into a (temp) layer. What to do when
there is a filter? This CL only passes through the filtered paint if it
was changed from the default state. This may still be too aggressive,
since the filter could have only changed aspects that don't affect
pictures...

Possible follow-on: only pass a paint if a drawPicture-specific aspect
changed:
- alpha, colorFilter, imageFilter, blendMode

Change-Id: I766c0a129d3bb2882cab976fb2780ef9a09278ee