Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> (2 comments)
> 
> This may sound dumb, but I think I don't understand why the CTM is provided by the SkMatrixProvider.
> 
> Is it logically possible to plumb this thing through in parallel to the CTM rather than to have it carry the CTM?  I'm getting spooked out by wondering what a provider _does_ w.r.t the CTM.  It feels like the CTM could change out from underneath you, which is spooky... and if it can't, is it just carried by the SkMatrixProvider for convenience?

Yeah, there are a couple things here:

Originally the matrix provider was only going to provide the other (named) matrices, and it would be in parallel to the CTM. Some of us thought that it seemed redundant, because the named matrices are all specifically local-to-<something>, and the CTM is one specific instance of that, so lets move that responsibility to the provider as well.

As it turns out... places in the stack where we change the CTM (often with a pre-concat matrix) should logically also be changing the named matrices in the same way, so that worked out nicely. In addition, when we start adding 3D primitives (vertices/atlas/patch being the first obvious choices), we'll need a version of the CTM that hasn't discarded Z. The provider lets us defer that decision to the specific SkDraw/GrOp.

Patch-set: 7
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0