Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I realized that even if the type went all the way to carrying const non-reference fields, you could still always just placement-new on top of it to change them... you can never really trust they'll not be changed.  Maybe we should just ignore everything I said yesterday?  I think I can probably live with "you'll get used to it".

That was kind of my thinking, but I wanted to put this together to see if it made things any better. Ultimately, we were always passing around const& SkMatrix before, and that was a reference to a non-const matrix on the device. Having it live inside the provider makes it feel a little different, because the provider itself is being passed around, though always as a const& as well. I think the changes in PS8 on the other CL (to get rid of setPostMatrix) are good - but this change might be unnecessary?

Patch-set: 1
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>