Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I see a 20% speedup in cpu processing using the global cache (hacked w/ a mutex)
> 
> Still plenty of time in OT::BEint() -- why isn't this guy inlined?

Good question, which OT::BEInt is it? There should be at least four like BEInt<T,1> BEInt<T,2> BEInt<T,3> BEInt<T,4>. They are all forced unaligned so that may haven something to do with it, but they are in the header. I'm assuming this is with a clang build?

Patch-set: 1