Update patch set 2
Patch Set 2:

I take it this is for some type where you can't just makeArray() and loop afterwards?

Is it crazy to expose commonArrayAlloc() instead (maybe with a different name)?  Whatever you're trying to do, whether this or something else in the future, can always be done with commonArrayAlloc() and a loop afterwards.

Patch-set: 2
Reviewer: Gerrit User 5045 <5045@5086850b-e599-37d2-8ec8-ff16aeccbf02>