Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I take it this is for some type where you can't just makeArray() and loop afterwards?
> 
> Is it crazy to expose commonArrayAlloc() instead (maybe with a different name)?  Whatever you're trying to do, whether this or something else in the future, can always be done with commonArrayAlloc() and a loop afterwards.

It's true I can expose commonArrayAlloc(), but then the array would not be initialized, and the values in it could be in an invalid state. With makeInitializedArray() all the elements are valid. If I call commonArrayAlloc(), and then immediately destroy the arena, the destructors of each element will be called with invalid data.

Patch-set: 2
Reviewer: Gerrit User 5210 <5210@5086850b-e599-37d2-8ec8-ff16aeccbf02>