Update patch set 1
Patch Set 1: Code-Review+1 Commit-Queue+2

Well this looks like it is working. But the more I think about I have no idea how it is working. We explicitly dlopen libvulkan.so and then get the function pointers of there. From what I was told libvulkan.so was just a sym link to the lib you copied here. So my intuition would have told me that we wouldn't actually be using your copied library. Maybe libvulkan.so actually is a "real" library that itself loads the libGLES_mali library. And thus it is now finding the copied one.

Anyways don't question something that is working, so lgtm.

I would say just blacklist the one failing test, either in this CL or I can do it in a follow on CL, and file a bug to take a look at it. Leave that up to you if you want to do it here or not.

Patch-set: 1
Reviewer: Gerrit User 5036 <5036@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1
Label: Commit-Queue=+2