Update patch set 1
Patch Set 1: Code-Review+1

I had a similar CL with the addition of changing the the API to be clearer.
https://skia-review.googlesource.com/c/skia/+/289788

I changed mine to match your change. We can land this and the start the conversation about the API change in my CL.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1