Place transformed drawing data at source space (0,0)
The direct device remains fully positioned,
and will be addressed in a following CL.

This is a step towards making the vertex data
constant. The path data is positioned at draw
time instead of at creation time.

The GPU data is not constant at this time, but
when all the vertex data is placed at (0,0) the
data can be const.

Bug: skia:10251

Change-Id: I24991e1ba76ab77f06bf040d88de93bc31002b31