[canvaskit] Properly fix Safari (detect if WebGL2 is not available)
I believe we had incorrectly been using the wrong emscripten-provided
helper to make our WebGL contexts. Thus, the parameter that was
specifying webGL 1 vs 2 was not getting properly passed in
(I believe because of a closure-induced problem in minification).

I validated this build on my local Mac+Safari test harness.

Bug: skia:10171
Change-Id: Ifd5e55a6b64407bf84168e118d8dda4738ef5487