Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> While looking at using CustomTF, I found the fixed builder glyph count somewhat annoying.
> 
> Also, I'm guessing the storage should eventually be a sparse data structure (map), since the glyph IDs returned from shaping are not necessarily sequential or starting at zero?  This nudges the public API in the sparse direction.

All known glyph formats have dense glyph ids that start at zero (since some codepoint to glyph map is going to be needed in any event). Also, there are users who depend on knowing how may glyphs there are in a font and that being an unchanging number. On the other hand, there's no reason for a font not to have a bunch of glyphs set to the same backing like empty or tofu or something. In fact that's how we use the font subsetter.

Patch-set: 2
Reviewer: Gerrit User 5140 <5140@5086850b-e599-37d2-8ec8-ff16aeccbf02>