Update patch set 3
Patch Set 3:

I withdraw the notion of only offering this in the new api if its awkward (though maybe we should design the new one so it isn't). If we can at least ensure that we can guard/conditionalize the new behavior in the old API, I'll be fine -- again, just to be ready if we find different images in old apps.

Patch-set: 3
Tag: mailMessageId=<CAERTzqxYx=TeW3_9V6jJuNhRLTdsRKn_kq4EAJrJNsVs+2BWbw@mail.gmail.com>