Update patch set 4
Patch Set 4: Commit-Queue+1

Guarded drawVertices to keep the old behavior on Android framework. (Android doesn't expose drawPatch). I was able to put the legacy behavior in canvas, making it somewhat obvious how to move that to Android's code, if we'd like.

re: Making it easier to do this with the newer vertices variants -- I think no matter how you slice it, that ends up still being cumbersome for the user. The whole point of this tweak is that the user just wants to draw a thing that samples a shader based on local position. Even if we did allow SkSL to bypass the old behavior, the SkSL you'd supply would be:

  in shader child;
  void main(float2 p, inout half4 color) {
    color = sample(p, child);
  }

... the extra level of indirection you'd be adding with the SkSL would be pure overhead. And if the above worked (with no custom data), then the problem is already solved -- don't bother making an SkSL shader on the paint that's just going to call another shader -- put the shader you actually want on the paint.

Patch-set: 4
Reviewer: Gerrit User 5631 <5631@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=+1