Fix rendering of very-thin dashed lines when MSAA is enabled.
The existing code would increase the rendered stroke width to 1px (in
screen-space) to ensure proper subpixel coverage when the dashed line to
render was very thin. However, the MSAA path relies on hardware MSAA to
subpixel coverage. The stroke width is now left as-is when MSAA is on.

Additionally, this CL includes some minor polish to coverage-related
code that I made while hunting for the root cause of the bug.

Bug: skia:10240
Change-Id: Ied685822b1d2c5c985c66b1c958954883e6671a9