Track render target dimensions in GrScissorState
This helps restructure code that frequently had to test for intersection
with the render target before setting the scissor. In another CL, it will
be used to properly apply the scissor test for approx-fit render targets
with stencil operations.

It also makes the clear ops and the internal clear functions read a little
 easier IMO.

Change-Id: I75671c9cc921f4696b1dd5231e02486090aa4282