Track render target dimensions in GrScissorState
This helps restructure code that frequently had to test for intersection
with the render target before setting the scissor. Additionally, now the
clear ops bounds match the backing store dimensions when they are full
screen.

For stencil clears, this means that a scissor equal to the logical size
of the render target remains scissored. We explicitly allow a color
clear with a scissor == render target dims to expand to the actual
backing store dimensions. But now that is reflected in the bounds of the
op.

A subsequent CL will apply this same reasoning to added draw ops,
ensuring that the RTC maintains the stencil buffer properly outside its
logical dimensions.

Change-Id: I75671c9cc921f4696b1dd5231e02486090aa4282