Use pow2 rectanization for the hw tessellation atlas
This particular atlas is only used for small-area paths, which means
at least one dimension of every path is guaranteed to be quite small.
So we use a pow2 atlas and transpose tall paths. Transposing makes it
so every path has a small height, which lends very well to efficient
pow2 atlas packing.

Change-Id: I409283c392117d17937526ab512bbfa290a28577