Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> CC Michael for the pow2 atlas/transpose thoughts. Maybe we can come up with a good way of merging pow2 atlases

I don't have many thoughts yet for how this might apply to the future clip atlas. That doesn't have any guarantee on the dimensions of the clip paths. Since the clip atlas only needs to store the intersection of the clip path with the draw bounds, it'd be interesting to explore trade offs of using the union of all draw bounds to create one atlas entry (the way CCPR does it), or by giving each draw+clip intersection its own entry. The scenario where I'm imagining that would be useful is if there's a large clip path and then small draws that intersect its boundary; rasterizing the entire clip path would be a waste. 

Once we can start working on a general clip atlas, keeping in mind the transpose packing makes sense. Would skyline benefit from the same thing, however? If the clip atlas only used the pow2 as a conservative decider on when to end the ops task, but used transpose to get more efficient packing, we'd need to make sure that skyline would still benefit or we'd lose the conservative property of the pow2 packer.

Patch-set: 3