Add make-continuous-headless target (headless option) to Makefile.
One immediate use-case for this is running tests when SSH'd into a
remote machine.  In theory, X11 forwarding could be used to forward the
remote Chrome UI.  In practice, Chrome is difficult to forward
correctly, and the options for doing so are empirically quite slow.  In
testing, forwarding from Ubuntu to Mac required passing
--use-gl=swiftshader to Chromium, and rendering was slow enough to
timeout tests.

Bug: NONE
Change-Id: Ibfaa406a73d293ea212d7983d487ea0de9722da5