Fix GM 'strict_constraint_no_red_allowed' by removing AA optimization.
GrTextureOps was attempting to detect subset-rectangles that wouldn't
affect the rendering output and could be ignored. Unfortunately, the
'strict_constraint_no_red_allowed' test found several cases where this
optimization attempt failed in subtle ways--small, one-pixel cracks on
the edge of the border when AA was off, and highly-visible red fuzz on
the edges of textures when MSAA was enabled.

This CL also cleans up the test code slightly.

Change-Id: I200aaab47737b5ba0f559182ef4d0dfe0b719d50