Update patch set 2
Patch Set 2: Code-Review+1

> Patch Set 1:
> 
> Hi folks, a couple of things:
> 
> 1. I was hoping I could get some assistance with some of the Infrastructure Playbook steps outlined here: https://cs.chromium.org/chromium/src/third_party/skia/modules/canvaskit/README.md?l=29-64&rcl=87caae61cdf1d59f393d975fb3c7fe5ea27f6206 .

Done. I uploaded PS 2 to fix a few typos in those docs, as well as wordsmith a few of the CHANGELOG.MD, but other than that, your change covered everything.

> 2. I did want to get your feedback on the breaking change that replaces `ready().then()` with `then()`.  I'm not sure there's a great way to maintain backwards compatibility here, or how relatively important that is for CanvasKit currently.

We follow semantic versioning, which I'll paraphrase "0.X.Y versions can change the API in dramatic ways." We follow the guideline of updating X to X+1 when we have breaking changes, and that's what we'll do here.

> 3. I also wanted to get your thoughts on streamlining updating emscripten in future.  Perhaps we can model this with DEPS or similar to have a emscripten version locked in a single place across all usages?
We have 3 types of places that need to get updated - the docker images, the Makefiles, and the recipes. I'm not sure of a great solution to unite all 3, but even if we can unite two of them or reduce the manual touch points, I'm all ears for ideas.

I triggered the tryjobs - if those pass, I'll submit.

Patch-set: 2
Reviewer: Gerrit User 5395 <5395@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1