Update patch set 3
Patch Set 3:

+Michael.

I'm afraid of the performance regressions, particularly in SkiaRenderer, that we may see if we land this. I'm also no longer 100% sure about the inputs to this texture op constructor mean. I'm wondering if maybe we should be doing a comparison of subset and quad->fLocal to see if the represent the same rectangle to make the optimization safe. Michael definitely knows this area better than I do. Michael, will quad->fLocal exactly represent the original src rect for a drawImageRect? Seems like it would.

I'm also wondering if we want to treat MSAA and non-MSAA differently. In MSAA I think whether the optimization is theoretically safe or not is determined by whether the GPU implements centroid sampling or not. If it it implements centroid sampling then in theory (at least) all the fragment shader samples should fall within the convex hull of the texture coordinates computed at the quad vertices. If it doesn't use centroid sampling then the samples are taken by interpolating the texture coords to the center of pixels that have covered samples. A sample can be covered by the the quad even if the corresponding pixel center is not.

I don't off the top of my head know what GL (and other APIs) require for MSAA interpolation at different versions, but I believe at least in ES 3.0 we can modify our shaders to require centroid interpolation per varying. I actually just tried editing the GLSL shader in viewer on my desktop to specify "centroid" on the varying declaration in the FS and VS with --msaa 8 and it seems to work! So perhaps all GL versions default to pixel center and then at some version allow you to override it with centroid.

Patch-set: 3
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Reviewer: Gerrit User 5085 <5085@5086850b-e599-37d2-8ec8-ff16aeccbf02>