Iterate over unresolved instead of cycling.
This makes it much clearer which code is actually adding to and removing
from the unresolved queue. This also makes fillGaps much more
performant since it no longer needs to make a full copy of the
unresolved blocks.

Change-Id: I62a5eb32118fec6745b7079f537ccbd07b018c12

It's so good to have a fresh pair of eyes...
I don't even notice these things anymore :(

Change-Id: I62a5eb32118fec6745b7079f537ccbd07b018c12