Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> +Michael.
> 
> I'm afraid of the performance regressions, particularly in SkiaRenderer, that we may see if we land this. I'm also no longer 100% sure about the inputs to this texture op constructor mean. I'm wondering if maybe we should be doing a comparison of subset and quad->fLocal to see if the represent the same rectangle to make the optimization safe. Michael definitely knows this area better than I do. Michael, will quad->fLocal exactly represent the original src rect for a drawImageRect? Seems like it would.
> 

It may hurt us here, but I don't think it would be so bad. As far as I remember, the times we needed strict constraint, it was definitely important that we got it right since the extra texels could have garbage data in them. Given the recent fix you made for insetting the subset when we have kNearest, I'm wondering if the premise of this optimization was entirely invalid? Replacing it with something that compares the local quad's bounds with the subset would be more correct.

In the case of drawImageRect, quad->fLocal will correspond to the original src rect.

Patch-set: 3
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>