Update patch set 11
Patch Set 11:

(1 comment)

> Patch Set 11:
> 
> I think I still have a few unanswered questions around this change. The choice to use window rectangles in the past was quite data driven, and showed a large improvement on mixed samples running the Chrome skps. Chrome would often had clips that included one very large difference rect, and I can imagine a clipping scenario like that where a window rectangle would disable most the clear. Mixed samples makes this even worse, because if you have 8 stencil samples and only 1 color sample, the relative cost of a stencil clear is quite high.
> 
> Do we have perf data with mixed samples? I thought we had some with MSAA that showed the window rectangle clears were helpful?
> 
> I'm also not sure I understand why this is needed for the next clipping CLs. I know we talked about refactoring GrFixedClip to be nothing but scissor + window rectangles. Would that work instead?
> 
> Maybe a good torture case to test would be a difference clip of a near-fullscreen, complex round rect with small arcs. Does analytic aa work on complex round rects now? If not, then window rectangles would erase most of it for the clear and I imagine we would see a perf win from using mixed samples in the clear.

I confirmed that window rectangles have a significant positive impact on performance, particularly for MSAA targets. This change does make it so that a native glClear call, or backend equivalent, won't use window rectangles. But the GrRTC functions detect this and do a non-AA axis-aligned rect draw with window rectangles instead of a glClear call. In my performance tests, this was as fast (or somehow sometimes even faster than a glClear).

Given that, and the 25-50% of backends that can support window rectangles, I wanted to isolate the places in the lower-level code that had to handle them. With this change, only drawing ops need to know about window rectangles because they are still stored in the GrAppliedHardClip object. Clear implementations can assume scissor or fullscreen (depending on the caps bit they set).

Regarding GrFixedClip being used as this vehicle for scissor state and window rectangle state: I didn't want the low level GPU code to deal with it as-is because it extends GrClip, which I am transitioning to be just a high-level clip operator that doesn't get past the GrRTC. Since there was only one place that actually performs a stencil clear with window rectangles, and doing it as a draw is straight forward and no cost to performance, I figured it was simpler to have it handle the window rectangles than introduce additional types into the clipping API.

Patch-set: 11
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Commit-Queue=0