Improve 'strict_constraint_no_red_allowed' by checking axis alignment.
GrTextureOps was attempting to detect subset-rectangles that wouldn't
affect the rendering output and could be ignored. Unfortunately, the
'strict_constraint_no_red_allowed' test found several cases where this
optimization attempt failed in subtle ways--small, one-pixel cracks on
the edge of the border when AA was off, and highly-visible red fuzz on
the edges of textures when MSAA was enabled. This CL limits the
optimization to cases where the source and destination quads are
axis-aligned rectangles, which fixes most of these artifacts.

When MSAA is enabled, there are still some small artifacts visible when
quads are not on pixel boundaries despite being axis-aligned. This
should be addressed in a followup CL.

This CL also cleans up the test code slightly.

Bug: skia:10263
Change-Id: I200aaab47737b5ba0f559182ef4d0dfe0b719d50