Update patch set 5
Patch Set 5:

> Patch Set 3:
>
> I subpixel offset is only a concern for MSAA, though. And if we can get centroid sampling that in theory make it work for MSAA, though we'd have to see how that pans out in practice. MSAA is the rare case. Most clients don't use MSAA at all (e.g. Android framework, Flutter). Chrome compositing never uses it. Chrome raster occasionally uses it.

I've brought back the optimization, but restricted it to cases where both the source and destination draw quads are axis-aligned. This doesn't solve the MSAA subpixel-boundary issue, but seems to resolve everything else nicely, and hopefully doesn't impact performance too terribly. (IMO we should address MSAA in a separate CL, since we have multiple potential avenues of attack--we can go for the centroid varyings, or disable the optimization when subpixel boundaries are involved, or disable the optimization when MSAA is enabled.)

I still want to go over Michael's list of related places in the code and fix those up as well before submitting this. 

(I also think I should probably try to put together a clone of the existing test which checks strict clamping but leverages the bulk-draw path, since apparently we don't have test coverage here. That would be for a separate CL though.)

Patch-set: 5
Reviewer: Gerrit User 11059 <11059@5086850b-e599-37d2-8ec8-ff16aeccbf02>