Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> The fix definitely seems like the right direction. Michael, should we check if quad->fLocal is the same as *subset rather than just that it is axis-aligned?

I think the best action would be to check if quad->fLocal.bounds() is contained within the .5px inset of the subsetRect. If that's the case, we should be able to drop it. I think this would then work for non-axis-aligned quads and even for bilerp? We could even move this into the normalize_subset function, but we'd have to then provide it the domain bounds and give it a way to report back that a subset is not needed.

With just checking that the type is axis-aligned, we'd discard a subset rectangle that would definitely clamp the local coords.  Checking that local is equal to the subset rect would also guard this, but contains would generalize it.

Patch-set: 5
Reviewer: Gerrit User 8341 <8341@5086850b-e599-37d2-8ec8-ff16aeccbf02>