Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> This new test is running several bots over memory.  I don't want to revert it given the attached bug, which I assume is serious.  Does it make sense to disable this unit test on those bots, e.g.
> 
>   Test-Android-Clang-Nexus7-GPU-Tegra3-arm-Debug-All-Android
>   Test-iOS-Clang-iPhone6-GPU-PowerVRGX6450-arm64-Debug-All
> 
> ?

I am trying to contact the ochang@google.com the security person who posted this bug. I do not think this is serious at all. The bug has been in the code since the first check-in five years ago. Chrome's html layout is protecting Skia from getting strings over 23M characters long, which is the breaking point for 32-bit machines.

I would be ok removing the tests, or #if define'ing them to only work on 32-bit platforms with large memory.

Thanks for tracking this down, and the advise.

Patch-set: 7