Update patch set 9
Patch Set 9: Code-Review+1

(2 comments)

Does the "auto" nature of this seem natural, and if so, how far down? I love the feature in general (orienting based on the tangent to the animation curve), but the auto part (deciding to orient or not) just feels slippery. It is being resolved at the right level?

Patch-set: 9
Reviewer: Gerrit User 5649 <5649@5086850b-e599-37d2-8ec8-ff16aeccbf02>
Label: Code-Review=+1