Update patch set 5
Patch Set 5:

I've rewritten this GM to draw a yellow circle with a blue square on top of a checkerboard. All blendmodes are tested by drawing the blue square with different blendmodes.

I believe this GM is more human readable now. 

The usage of the checkerboard in this GM is different than in the GMs Brian linked in the previous comment. In those examples the checkerboard represents transparency, so the circle and square blendmode testing image is prepared separately before drawing it on top of a checkerboard. 
In this CL's GM the shapes are prepared on top of the checkerboard so the blendmodes affect the checkerboard as well. 

As such, the appearance of this GM may not match expectations for a GM using a checkerboard. I would like to avoid this problem, but I do not know a simple way to prepare the shape image separately using just the SkCanvas API in canvas.spec.ts.

Patch-set: 5