Avoid malloc calls for these two temp arrays on the stack.
Definitely improves perf when I perform lots of layouts.

Don't have a strong opinion precisely how to do it. Perhaps using a stack-hosted SkArenaAlloc would work too.

TODO: fRunsInVisualOrder is a long-lived array, but appears to..
- never resize after the constructor
- very often be size==1
Can we preallocate storage for it in the TextLine itself? (e.g. StSTArray or other trick)

Change-Id: I817b46a24e01ddf999bdd81a607aaf35b3c0674b