Update patch set 1
Patch Set 1:

It looks like I am still getting my compiler error. I'm not sure if the intent of this particular CL was to fix that, but you can test it by running:

  bin/fetch-skps. # Only need to do this once

  ninja -C out/debug/ viewer && out/debug/viewer -b mtl --msaa 8 --pr tess --internalSamples 8 --match chalkboard

Note that Metal might still crash after the fix, but we at least shouldn't be getting the compiler error printed out to the terminal.

Patch-set: 1